+++ packages/config/CHANGELOG.md
@@ -102,3 +102,86 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 - **config:** levels now should be arrays ([ef142d8](https://github.com/bem/bem-sdk/commit/ef142d8))
+
+## Pre-1.0 history (legacy)
+
+# [0.1.0](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.10...@bem/sdk.config@0.1.0) (2019-04-15)
+
+
+### Features
+
+* **config:** merge common opts to each level ([349460a](https://github.com/bem/bem-sdk/commit/349460a))
+
+
+
+
+
+<a name="0.0.10"></a>
+## [0.0.10](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.9...@bem/sdk.config@0.0.10) (2018-07-01)
+
+
+
+
+**Note:** Version bump only for package @bem/sdk.config
+
+<a name="0.0.9"></a>
+## [0.0.9](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.7...@bem/sdk.config@0.0.9) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package @bem/sdk.config
+
+<a name="0.0.7"></a>
+## [0.0.7](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.6...@bem/sdk.config@0.0.7) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package @bem/sdk.config
+
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.5...@bem/sdk.config@0.0.6) (2017-12-27)
+
+
+### Bug Fixes
+
+* **config:** no need to dynamically load plugins ([9eb8df2](https://github.com/bem/bem-sdk/commit/9eb8df2))
+
+
+
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.4...@bem/sdk.config@0.0.5) (2017-12-12)
+
+
+
+
+**Note:** Version bump only for package @bem/sdk.config
+
+<a name="0.0.4"></a>
+## [0.0.4](https://github.com/bem/bem-sdk/compare/@bem/sdk.config@0.0.3...@bem/sdk.config@0.0.4) (2017-11-07)
+
+
+
+
+**Note:** Version bump only for package @bem/sdk.config
+
+<a name="0.0.3"></a>
+## 0.0.3 (2017-10-01)
+
+
+### Bug Fixes
+
+* **config:** levels now should be arrays ([ef142d8](https://github.com/bem/bem-sdk/commit/ef142d8))
+
+
+
+
+<a name="0.0.2"></a>
+## 0.0.2 (2017-09-30)
+
+
+### Bug Fixes
+
+* **config:** levels now should be arrays ([ef142d8](https://github.com/bem/bem-sdk/commit/ef142d8))